+++ AIO_Firmware_PIO/src/app/weather/weather.cpp
@@ -578,7 +578,7 @@ static void weather_message_handle(const char *from, const char *to,
     case APP_MESSAGE_WIFI_CONN:
     {
         Serial.println(F("----->weather_event_notification"));
-        int event_id = (int)message;
+        int event_id = (int)(intptr_t)message;
         switch (event_id)
         {
         case UPDATE_NOW:
